+++ src/Symfony/Component/Messenger/Tests/Command/ConsumeMessagesCommandTest.php
@@ -22,7 +22,6 @@ use Symfony\Component\Messenger\Command\ConsumeMessagesCommand;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\RoutableMessageBus;
-use Symfony\Component\Messenger\Stamp\BusNameStamp;
 use Symfony\Component\Messenger\Transport\Receiver\ReceiverInterface;
 
 class ConsumeMessagesCommandTest extends TestCase
@@ -37,38 +36,6 @@ class ConsumeMessagesCommandTest extends TestCase
 
     public function testBasicRun()
     {
-        $envelope = new Envelope(new \stdClass(), [new BusNameStamp('dummy-bus')]);
-
-        $receiver = $this->createMock(ReceiverInterface::class);
-        $receiver->expects($this->once())->method('get')->willReturn([$envelope]);
-
-        $receiverLocator = $this->createMock(ContainerInterface::class);
-        $receiverLocator->expects($this->once())->method('has')->with('dummy-receiver')->willReturn(true);
-        $receiverLocator->expects($this->once())->method('get')->with('dummy-receiver')->willReturn($receiver);
-
-        $bus = $this->createMock(MessageBusInterface::class);
-        $bus->expects($this->once())->method('dispatch');
-
-        $busLocator = $this->createMock(ContainerInterface::class);
-        $busLocator->expects($this->once())->method('has')->with('dummy-bus')->willReturn(true);
-        $busLocator->expects($this->once())->method('get')->with('dummy-bus')->willReturn($bus);
-
-        $command = new ConsumeMessagesCommand(new RoutableMessageBus($busLocator), $receiverLocator, new EventDispatcher());
-
-        $application = new Application();
-        $application->add($command);
-        $tester = new CommandTester($application->get('messenger:consume'));
-        $tester->execute([
-            'receivers' => ['dummy-receiver'],
-            '--limit' => 1,
-        ]);
-
-        $this->assertSame(0, $tester->getStatusCode());
-        $this->assertStringContainsString('[OK] Consuming messages from transports "dummy-receiver"', $tester->getDisplay());
-    }
-
-    public function testRunWithBusOption()
-    {
         $envelope = new Envelope(new \stdClass());
 
         $receiver = $this->createMock(ReceiverInterface::class);
@@ -82,8 +49,8 @@ class ConsumeMessagesCommandTest extends TestCase
         $bus->expects($this->once())->method('dispatch');
 
         $busLocator = $this->createMock(ContainerInterface::class);
-        $busLocator->expects($this->once())->method('has')->with('dummy-bus')->willReturn(true);
-        $busLocator->expects($this->once())->method('get')->with('dummy-bus')->willReturn($bus);
+        $busLocator->expects($this->once())->method('has')->with('dummy-receiver')->willReturn(true);
+        $busLocator->expects($this->once())->method('get')->with('dummy-receiver')->willReturn($bus);
 
         $command = new ConsumeMessagesCommand(new RoutableMessageBus($busLocator), $receiverLocator, new EventDispatcher());
 
@@ -92,80 +59,6 @@ class ConsumeMessagesCommandTest extends TestCase
         $tester = new CommandTester($application->get('messenger:consume'));
         $tester->execute([
             'receivers' => ['dummy-receiver'],
-            '--bus' => 'dummy-bus',
-            '--limit' => 1,
-        ]);
-
-        $this->assertSame(0, $tester->getStatusCode());
-        $this->assertStringContainsString('[OK] Consuming messages from transports "dummy-receiver"', $tester->getDisplay());
-    }
-
-    /**
-     * @group legacy
-     * @expectedDeprecation Passing a "Psr\Container\ContainerInterface" instance as first argument to "Symfony\Component\Messenger\Command\ConsumeMessagesCommand::__construct()" is deprecated since Symfony 4.4, pass a "Symfony\Component\Messenger\RoutableMessageBus" instance instead.
-     */
-    public function testBasicRunWithBusLocator()
-    {
-        $envelope = new Envelope(new \stdClass(), [new BusNameStamp('dummy-bus')]);
-
-        $receiver = $this->createMock(ReceiverInterface::class);
-        $receiver->expects($this->once())->method('get')->willReturn([$envelope]);
-
-        $receiverLocator = $this->createMock(ContainerInterface::class);
-        $receiverLocator->expects($this->once())->method('has')->with('dummy-receiver')->willReturn(true);
-        $receiverLocator->expects($this->once())->method('get')->with('dummy-receiver')->willReturn($receiver);
-
-        $bus = $this->createMock(MessageBusInterface::class);
-        $bus->expects($this->once())->method('dispatch');
-
-        $busLocator = $this->createMock(ContainerInterface::class);
-        $busLocator->expects($this->once())->method('has')->with('dummy-bus')->willReturn(true);
-        $busLocator->expects($this->once())->method('get')->with('dummy-bus')->willReturn($bus);
-
-        $command = new ConsumeMessagesCommand($busLocator, $receiverLocator, new EventDispatcher());
-
-        $application = new Application();
-        $application->add($command);
-        $tester = new CommandTester($application->get('messenger:consume'));
-        $tester->execute([
-            'receivers' => ['dummy-receiver'],
-            '--limit' => 1,
-        ]);
-
-        $this->assertSame(0, $tester->getStatusCode());
-        $this->assertStringContainsString('[OK] Consuming messages from transports "dummy-receiver"', $tester->getDisplay());
-    }
-
-    /**
-     * @group legacy
-     * @expectedDeprecation Passing a "Psr\Container\ContainerInterface" instance as first argument to "Symfony\Component\Messenger\Command\ConsumeMessagesCommand::__construct()" is deprecated since Symfony 4.4, pass a "Symfony\Component\Messenger\RoutableMessageBus" instance instead.
-     */
-    public function testRunWithBusOptionAndBusLocator()
-    {
-        $envelope = new Envelope(new \stdClass());
-
-        $receiver = $this->createMock(ReceiverInterface::class);
-        $receiver->expects($this->once())->method('get')->willReturn([$envelope]);
-
-        $receiverLocator = $this->createMock(ContainerInterface::class);
-        $receiverLocator->expects($this->once())->method('has')->with('dummy-receiver')->willReturn(true);
-        $receiverLocator->expects($this->once())->method('get')->with('dummy-receiver')->willReturn($receiver);
-
-        $bus = $this->createMock(MessageBusInterface::class);
-        $bus->expects($this->once())->method('dispatch');
-
-        $busLocator = $this->createMock(ContainerInterface::class);
-        $busLocator->expects($this->once())->method('has')->with('dummy-bus')->willReturn(true);
-        $busLocator->expects($this->once())->method('get')->with('dummy-bus')->willReturn($bus);
-
-        $command = new ConsumeMessagesCommand($busLocator, $receiverLocator, new EventDispatcher());
-
-        $application = new Application();
-        $application->add($command);
-        $tester = new CommandTester($application->get('messenger:consume'));
-        $tester->execute([
-            'receivers' => ['dummy-receiver'],
-            '--bus' => 'dummy-bus',
             '--limit' => 1,
         ]);